04_interactive_agent/pyproject.toml
[project]
name = "04-interactive-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk[http-server]>=0.3.5",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.10",
    "langgraph>=0.6.7",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
